backend/requirements.txt
# Backend requirements
fastapi
python-multipart
uvicorn[standard]
pydantic>=2
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
openai
slowapi


